Reuse is_coroutine_callable
